Move ServerPlayerManager back to Players namespace
